pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "startup-edu-agent"
version = "0.1.0"
description = "MVP for OCR-backed startup coaching with DeepSeek integration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1.0",
  "openai>=1.51.0,<2.0",
  "pydantic>=2.8,<3.0",
  "pymupdf>=1.24.0,<2.0",
  "pyyaml>=6.0,<7.0",
  "python-multipart>=0.0.9,<1.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
ocr = [
  "pillow>=10.4,<12.0",
  "pytesseract>=0.3.13,<1.0",
]
test = [
  "pytest>=8.3,<9.0",
]
ui = [
  "streamlit>=1.41,<2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
